+++ authentication/admin.py
@@ -1,8 +1,6 @@
 from django.contrib import admin
 
 # Register your models here.
-from authentication.models import CustomUser, Card, Resource
+from authentication.models import CustomUser
 
 admin.site.register(CustomUser)
-admin.site.register(Card)
-admin.site.register(Resource)
